Create README.md
Create an image cloud based on a text file. I love Python graphics and this was to show how to create an image based on the contents of a text file. It can have a title printed on it and also have a signature. This example has the text image made up of my Python modules Have fun and play hard!